+++ test/collection_test.dart
@@ -2267,6 +2267,36 @@ void main() {
         expect('   3\n  2\n 1'.dedent(ignoreEmpty: false), '  3\n 2\n1');
       });
     });
+    group('take/skip', () {
+      test('take', () {
+        expect('abc'.take(0),'');
+        expect('abc'.take(1),'a');
+        expect('abc'.take(2),'ab');
+        expect('abc'.take(3),'abc');
+        expect('abc'.take(4),'abc');
+      });
+      test('skip', () {
+        expect('abc'.skip(0),'abc');
+        expect('abc'.skip(1),'bc');
+        expect('abc'.skip(2),'c');
+        expect('abc'.skip(3),'');
+        expect('abc'.skip(4),'');
+      });
+      test('takeLast', () {
+        expect('abc'.takeLast(0),'');
+        expect('abc'.takeLast(1),'c');
+        expect('abc'.takeLast(2),'bc');
+        expect('abc'.takeLast(3),'abc');
+        expect('abc'.takeLast(4),'abc');
+      });
+      test('skipLast', () {
+        expect('abc'.skipLast(0),'abc');
+        expect('abc'.skipLast(1),'ab');
+        expect('abc'.skipLast(2),'a');
+        expect('abc'.skipLast(3),'');
+        expect('abc'.skipLast(4),'');
+      });
+    });
     group('wrap', () {
       test('default', () {
         expect('a'.wrap(4), 'a');